Move pkg/bearertoken to internal/bearertoken.
