+++ cmd/remote-storage/app/server.go
@@ -23,7 +23,6 @@ import (
 	"github.com/jaegertracing/jaeger/internal/storage/v1/grpc/shared"
 	"github.com/jaegertracing/jaeger/internal/telemetry"
 	"github.com/jaegertracing/jaeger/internal/tenancy"
-	"github.com/jaegertracing/jaeger/pkg/telemetry"
 )
 
 // Server runs a gRPC server
