+++ internal/storage/metricstore/prometheus/metricstore/reader_test.go
@@ -28,9 +28,9 @@ import (
 
 	"github.com/jaegertracing/jaeger/internal/bearertoken"
 	config "github.com/jaegertracing/jaeger/internal/config/promcfg"
+	"github.com/jaegertracing/jaeger/internal/prometheus/config"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/metricstore"
 	"github.com/jaegertracing/jaeger/internal/testutils"
-	"github.com/jaegertracing/jaeger/pkg/prometheus/config"
 	"github.com/jaegertracing/jaeger/proto-gen/api_v2/metrics"
 )
